pyproject.toml
[project]
name = "azure-rag-playground"
version = "1.0.0"
description = "The app allows users to input a question and retrieves relevant information from Azure Search to provide an answer using OpenAI's language model."
authors = [{ name = "Chaitanya Madduri" }]
requires-python = ">=3.12"

dependencies = [
    "httpx==0.27.2",
    "typer==0.15.2",
    "openai>=1.83.0",
    "streamlit==1.42.0",
    "watchdog==4.0.0",
    "python-dotenv>=1.1.0",
    "ipykernel",
    "langchain-text-splitters",
    "pypdf",
    "sentence-transformers",
    "faiss-cpu",
    "redis"
    # If you have a GPU, replace "faiss-cpu" with "faiss"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
